requirements.txt
# Core
telethon>=1.34.0
python-dotenv>=1.0.0

# Supabase
supabase>=2.0.0

# Development
ruff>=0.8.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
